requirements.txt
crewai
python-dotenv
streamlit
litellm